+++ src/components/common/TagForm/index.tsx
@@ -1,0 +1,2 @@
+export type { TagFormProps } from './TagForm';
+export { default } from './TagForm';
