+++ src/pages/tag/index.tsx
@@ -5,7 +5,7 @@ import { GhostButton } from '~/components/common/Button';
 import LoadingHandler from '~/components/common/LoadingHandler';
 import NavigationBar from '~/components/common/NavigationBar';
 import { FixedSpinner } from '~/components/common/Spinner';
-import TagForm from '~/components/TagForm';
+import TagForm from '~/components/common/TagForm';
 import { defaultFadeInVariants } from '~/constants/motions';
 import useGetTagListWithInfinite from '~/hooks/api/tag/useGetTagListWithInfinite';
 import useInternalRouter from '~/hooks/common/useInternalRouter';
